ROpenSfM/requirements.txt
cloudpickle==0.4.0
exifread==2.1.2
fpdf2==2.0.6 #2.1.0
joblib==0.14.1
matplotlib
networkx==1.11
numpy
Pillow  #==7.1.0
pyproj>=1.9.5.1
pytest==3.0.7
python-dateutil==2.6.0
pyyaml==5.1
scipy
six
xmltodict==0.10.2
wheel
